+++ accounts/templates/accounts/participant_list.html
@@ -16,14 +16,8 @@
         });
     </script>
     <div class="row">
-        <div class="col-sm-10 col-lg-8 col-xs-12">
+        <div class="col-xs-12">
             <h1> Participants </h1>
-        </div>
-
-        <div class="col-xs-12 col-sm-2 pt-sm pull-right">
-            <a class="btn btn-primary" href="{% url 'exp:participant-email' %}">
-            <i class="fa fa-envelope-o" aria-hidden="true"></i> Email Participants
-            </a>
         </div>
     </div>
     <div class="row text-center mt-xl">
