+++ exp/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 '''
 from django.conf.urls import url
 
-from exp.views import (ParticipantDetailView, ParticipantListView, ParticipantEmailView,
+from exp.views import (ParticipantDetailView, ParticipantListView, StudyParticipantEmailView,
                        PreviewProxyView, ResearcherDetailView,
                        ResearcherListView, StudyBuildView, StudyCreateView,
                        StudyDetailView, StudyListView, StudyResponsesList,
@@ -27,10 +27,10 @@ urlpatterns = [
     url(r'researchers/(?P<pk>\d+)/$', ResearcherDetailView.as_view(), name='researcher-detail'),
     url(r'participants/$', ParticipantListView.as_view(), name='participant-list'),
     url(r'participants/(?P<pk>\d+)/$', ParticipantDetailView.as_view(), name='participant-detail'),
-    url(r'participants/email/$', ParticipantEmailView.as_view(), name='participant-email'),
     url(r'studies/$', StudyListView.as_view(), name='study-list'),
     url(r'studies/create/$', StudyCreateView.as_view(), name='study-create'),
     url(r'studies/(?P<pk>\d+)/$', StudyDetailView.as_view(), name='study-detail'),
+    url(r'studies/(?P<pk>\d+)/email/$', StudyParticipantEmailView.as_view(), name='study-participant-email'),
     url(r'studies/(?P<pk>\d+)/edit/$', StudyUpdateView.as_view(), name='study-edit'),
     url(r'studies/(?P<pk>\d+)/edit/build/$', StudyBuildView.as_view(), name='study-build'),
     url(r'studies/(?P<pk>\d+)/responses/$', StudyResponsesList.as_view(), name='study-responses-list'),
