+++ studies/templates/studies/study_participant_email.html
@@ -23,14 +23,15 @@
     <div class='row'>
          <div class="col-xs-12">
              <ol class="breadcrumb">
-               <li><a href="{% url 'exp:participant-list' %}"> All Participants </a></li>
-               <li class="active"> Send Email </li>
+               <li><a href="{% url 'exp:study-list' %}"> Manage Studies </a></li>
+               <li><a href="{% url 'exp:study-detail' pk=study.id %}"> {{ study.name }} </a></li>
+               <li class="active"> Email Participants </li>
              </ol>
          </div>
      </div>
-     <div class="row">
-         <div class="col-xs-12 col-xs-offset-0 col-md-10 col-md-offset-1">
-             {% if object_list %}
+     {% if participants %}
+        <div class="row">
+             <div class="col-xs-12 col-xs-offset-0 col-md-10 col-md-offset-1">
                  <div class="panel panel-default">
                    <div class="panel-heading">
                      <h1 class="panel-title">Contact Form</h1>
@@ -45,7 +46,7 @@
                          <div class="form-group">
                              <label for="recipients">Recipients</label>
                              <select multiple class="form-control" id="recipients" required name="recipients">
-                                 {% for participant in object_list %}
+                                 {% for participant in participants %}
                                     <option value="{{participant.id}}">{{ participant.given_name }}, ID: {{ participant.id }}</option>
                                  {% endfor %}
                              </select>
@@ -60,7 +61,7 @@
                             <textarea class="form-control" rows="5" id="message" required name="message"></textarea>
                          </div>
                          {% buttons %}
-                             <a href="{% url 'exp:participant-email' %}" class="btn btn-default">
+                             <a href="{% url 'exp:study-participant-email' study.id %}" class="btn btn-default">
                                  <i class="fa fa-trash-o" aria-hidden="true"></i>
                              </a>
                              <div class="pull-right">
@@ -72,10 +73,14 @@
                      </form>
                  </div>
              </div>
-             {% else %}
-                <p> No eligible participants to email. </p>
-             {% endif %}
          </div>
      </div>
+     {% else %}
+        <div class="row">
+            <div class="col-xs-12">
+                <p class="pt-md"><em> No eligible participants to email. </em></p>
+            </div>
+        </div>
+    {% endif %}
 </div>
 {% endblock %}
